att-decode: extract per-opcode ATT helpers into att-decoders.ts
